# docs/swagger.yaml
definitions:
...
  models.City:
    properties:
      id:
        type: integer
      lat:
        type: string
      long:
        type: string
      name:
        type: string
    type: object
  models.Destination:
    properties:
      address:
        type: string
      category:
        type: string
      city:
        $ref: '#/definitions/models.City'
      city_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      facilities:
        type: string
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/models.Image'
        type: array
      name:
        type: string
      operational_hours:
        type: string
      position:
        type: number
      ticket_price:
        type: number
      video_contents:
        items:
          $ref: '#/definitions/models.VideoContent'
        type: array
    type: object
  models.Image:
    properties:
      destination_id:
        type: integer
      id:
        type: integer
      url:
        type: string
    type: object
  models.Route:
    properties:
      cost:
        type: integer
      created_at:
        type: string
      destinationCityName:
        type: string
      destinations:
        items:
          $ref: '#/definitions/models.RouteDestination'
        type: array
      distance:
        type: number
      id:
        type: integer
      originCityName:
        type: string
      time:
        type: string
      userID:
        type: integer
    type: object
  models.RouteDestination:
    properties:
      created_at:
        type: string
      destinationID:
        type: integer
      id:
        type: integer
      routeID:
        type: integer
    type: object
  models.VideoContent:
    properties:
      description:
        type: string
      destination_id:
        type: integer
      id:
        type: integer
      title:
        type: string
      url:
        type: string
    type: object
  request.CreateDestinationInput:
    properties:
      address:
        type: string
      category:
        type: string
      city:
        type: string
      description:
        type: string
      facilities:
        type: string
      image:
        items:
          type: string
        type: array
      name:
        type: string
      operational_hours:
        type: string
      position:
        type: number
      ticket_price:
        type: number
      video_contents:
        items:
          $ref: '#/definitions/request.VideoInput'
        type: array
    type: object
  request.CreateRouteInput:
    properties:
      cost:
        type: integer
      destinationCityName:
        type: string
      destinations:
        items:
          type: integer
        type: array
      distance:
        type: number
      originCityName:
        type: string
      time:
        type: string
      userID:
        type: integer
    type: object
  request.VideoInput:
    properties:
      description:
        type: string
      title:
        type: string
      url:
        type: string
    type: object
...
host: www.tripwise.my.id
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Documentation TripWise Capstone Project - Team 3
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TRIPWISE API
  version: "1.0"
paths:
...
  /destination:
    get:
      consumes:
      - application/json
      description: Fetch the destinations between two cities, based on the origin
        and destination city names
      parameters:
      - description: Origin city name
        in: query
        name: origin
        required: true
        type: string
      - description: Destination city name
        in: query
        name: destination
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: City not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to fetch destinations
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get destinations for a specific route
      tags:
      - Routes
    post:
      consumes:
      - application/json
      description: Create a new destination and associate it with a city, images,
        and video contents
      parameters:
      - description: Destination Input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.CreateDestinationInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Destination'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new destination
      tags:
      - Destinations
...
  /route:
    get:
      consumes:
      - application/json
      description: Fetch all routes created by a specific user
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all routes by user
      tags:
      - Routes
    post:
      consumes:
      - application/json
      description: Create a new route by specifying the origin and destination cities,
        and additional route details
      parameters:
      - description: Route details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.CreateRouteInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Route'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to create route
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new travel route
      tags:
      - Routes
  /route/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a route by its ID, including all related destinations
      parameters:
      - description: Route ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Route successfully deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid route ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Route not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to delete route
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a specific route
      tags:
      - Routes
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"